countme: Retry hourly on failure
This unit is triggered only once par week at a random time. Thus if it
fails, it will never retry for the week thus it won't be counted in the
corresponding Count Me window. The common transient failures are due to
the network not being available.

Add an hourly retry on failure to avoid this specific case. Couting will
still be performed at a random time but shifted by a couple of hours.
Counting will also only be done once as the logic for that check is
directly performed by rpm-ostree.

Fixes: https://github.com/coreos/rpm-ostree/issues/3033